Fix the recipe version bumping
The previous version was incorrect. It should not bump the version if
there wasn't a previous one, and the logic depends on reading an
existing recipe to compare the versions.

Fixed tests to reflect the correct behavior, and added new tests to make
sure a 2nd commit with the same version would bump the patch level.