Add /recipes/info handling
This is slightly different than the bdcs-api-rs version. It adds an
'errors' list used to return the reason for failing to get a recipe.

Because more than 1 recipes can be requested at a time the 3 lists may
or may not have the expected data in them. If a recipe was successfully
retrieved it will have an entry in the 'changes' and 'recipes' lists,
and if there was an error it will only have an entry in the 'errors'
list.