Change the way git commit times are converted.
Instead of using a GLib utility function to convert a time to a string,
navigating the various GLib time types along the way, just pull all of
the integral values out of the object we get from ggit and format the
string ourselves. This accomplishes a handful of things:

  - This removes a bunch of error checking we needed while converting
    from DateTime to TimeVal, which is kind of weird to have anyway
    because there's no good reason for any of that to fail.
  - This removes the gross side-effecty parts of converting to timeval
  - This handles an API change in g_time_val_to_iso8601 (return val is
    nullable since glib-2.53.2, also another potential error that should
    never be an error) by removing that call entirely.

Also remove the comments about time zone handling. Unless libgit2-glib
is doing something wrong, the time zone embedded in the DateTime we get
should be correct and convertible to UTC.