Fix setup of the temporary repo and expand the test changes and new
The withClient code is executed for each of the 'it' sections so this
was causing problems with the setup of the repo, it would be removed
before each test.

Extending the /recipes/new test to commit 10 more changes, and checking
that offset/limit of /recipes/changes works with a longer list of
changes is what revealed this problem.