Mount source dir inside Docker container
this commit alters how bdcs-api is built within the docker container.
Instead of compiling during build time we now compile during runtime
and bind-mount the local directory to the running container.

This allows us to configure Travis so it caches the local sandbox
which should speed up testing because we're not going to rebuild
all of the underlying Haskell dependencies every time.