Add the /compose route.
This works and generates a compose, but it's pretty rough at the moment.
The biggest problem with it is that it blocks.  The client will not get
a response with the build ID back until the compose is finished.
Obviously, that needs to be fixed before this is used for real.