Use a worker thread and queue to do composes without blocking.
This is really simple - the server creates a queue for communication and
then spawns a thread.  The queue contains a new ComposeInfo record which
contains everything needed to do a compose.  The worker thread pulls
things out of that queue one at a time, does a compose, and writes a
status file.

This could probably be adapted later to pull multiple things out of the
queue to do multiple simultaneous composes.  It could potentially also
be adapted to not write a status file and instead be asked directly for
status.

I'm not convinced that the SQL pool portion is right at the moment.  It
may in the future need some locking to prevent multiple simultaneous
accesses.  I don't know yet.