Add a way for the API server to ask things of the compose threads.
This is not yet in use, but will be soon.  I first need to make some
additional changes to the compose thread portion.

The idea here is that getting information about the status of the
queue, canceling running composes, etc. will be easier to do with some
sort of communications channel and messages.  Responses will not use
this same channel, since making a single TMVar bidirectional is weird
and error prone.