Rework the thread that does composes.
This got complicated in a hurry, but is necessary for being able to see
what's in the compose work queue and letting clients know what is going
on.

First, change the work queue to a list from a TQueue.  We can't see the
entirety of the latter because it's a queue, but we can if it's just a
list.  This is needed to implement the /compose/queue route.

Second, make the compose thread a lot more complicated.  It needs to
fork off each compose as yet another thread so we can later implement
/compose/cancel and potentially support multiple simultaneous composes.
An IORef is currently used to keep track of whatever compose is in
progress.

It also needs to both respond to messages from the API (which is needed
to implement /compose/queue and probably a lot more routes) and watch
the queue at the same time.