Move depsolveRecipe out into a library function.
This is needed so it can be called from the compose function.  This in
turn is needed because that function happens in a forked off thread,
therefore putting dependency solving into the thread too.  This results
in the client getting a UUID returned to it much faster.

I'm somewhat confident this results in the same depsolve results.