Add build information to the /projects/info route (#36).
Doing so first required adding a whole lot of tedious but not difficult
types that make up the server's response.  It then required a lot of
database reading code since we spread project, source, and build
information out all over the place.

Note that I'm pretty extensively using ExceptT here including in a
couple places where a MaybeT might be better.  The little fetchProjects
helper function is a great example of this.  The database code wants to
return an ExceptT, but it makes little sense for the helper to do so
since it really does just return a Maybe.  For now I'm leaving it alone.
A later commit might make this cleaner.

Also note that the metadata objects are not being written at the time -
they're just empty lists.  I think that's enough code to require a
separate commit, and I don't think it's strictly necessary to get the
bug fixed.