Fix building tests in jenkins.
The main problem here is that we install hlint (and hpc-coveralls and
happy) into the sandbox as one transaction, and then install the build
dependencies into the same sandbox as a second transaction.  The first
one may result in versions of things being installed that are
incompatible with the second.

The fix is to install these programs globally into the container,
meaning they are not installed in the sandbox and cannot interfere.  We
then just look for them in $PATH.

We also have to delete the sandbox before doing anything with it,
because the source directory gets mounted into the container.  If the
source directory already has a sandbox, weird build errors could result.
Just clean it up first.