Add required packages to the compose if they're not already specified.
This is probably not the complete list of packages, but it's a start.
Certain packages are required to be in the compose or else it will fail.
Ostree composes in particular run various programs that are assumed to
be in the compose.

They are added without any version numbers or anything.  It is assumed
if the user cared about specific versions, they would list them in the
blueprint.  Otherwise, they just get whatever makes the dependencies
happy.