Rework composeInfo to use ExceptT.
The idea here was that it would make error handling easier, which would
mean les sstairstepping.  I think it does that, but I'm not convinced
the result is all that much easier to read.  Getting all the types to
line up make this more verbose than it really should be.