Add the /compose/cancel route.
This introduces two new messages that the server can handle - one for
canceling composes that have not yet started, and one for canceling
composes that are in progress.  Performing these two actions is very
different, so different messages seems reasonable.

Note also that this changes the return type from what we provided in the
python API.  Instead of returning some type containing the UUID, I'm
returning an APIResponse.  This is in line with a lot of our other
routes, and it's assumed the client already knows the UUID since that
was passed in to do the cancel in the first place.