Update the compose server to handle multiple simultaneous composes.
Much of the change in this patch is just removing the use of a TVar to
handle synchronization of the worklist between threads, replacing it
with a TMVar.  The TMVar supports retry operations critical for making
this work.

The other big change is replacing composesThread.  That code had a
single thread sitting and watching the worklist, waiting for something
to show up.  That organization made it difficult to have many composes
running at the same time.

Instead, I fork off a number of worker threads immediately.  They all
sit around watching the worklist, waiting for a compose request to
appear.  When a request does appear, a random worker thread will grab it
first and do the compose.  This means I don't ever need to worry about
the size of the in progress map, because the max number of composes
running at a time is constrained by how many workers are spawned.  It
also means the workers are less complicated.

For now, I am leaving max composes at 1.  I have tested two running at a
time, but only on the management side.  I don't know if what's going
into the composes is thread safe.  We need to check that before raising
this number or making it tunable by end users.