Add a set of default enabled language extensions.
The extensions in this list in the cabal file will be enabled for all
source files, so I've only made it the ones that are both widely used
and don't do anything weird or potentially unsafe or very slow.  Others
can be added as needed.

This also means the LANGUAGE pragmas for these extensions can be removed
from the source files.